requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
pyodbc==5.1.0
python-dotenv==1.0.1
APScheduler==3.10.4
requests==2.32.3
wfastcgi==3.0.0
